% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.prepare.R
\name{app.prepare}
\alias{app.prepare}
\title{Plant Catalog of Units of Brazilian Conservation App}
\usage{
app.prepare()
}
\value{
CSV files
}
\description{
An R app for preparing species listings for the Plant Catalog of Units of Brazilian Conservation.
}
\details{
limpar memória

direcionar memória para processamento temporário em disco

carregar funcões para mensurar tempos de processamento

inicar tempo de processamento

carregar pacotes básicos

cerregar funções desenvolvidas pelo CNCFLora

baixar e tabela FB2020 IPT
Rodar somente em atualizações do IPT, aproximadamente 6 horas de processamento.

carregar tabela FB2020 IPT e funções de acesso
conferência taxonômica
carregar informações da espécie
somente aqui encontramos dados de tipo de vegetação conforme FB2020

carregar funções para acesso APIs FB2020 v1 e v2
conferência taxonômica
carregar informações da espécie

carregar pacotes R e funções

finalizar tempo de processamento

pano de fundo

Tela junção de resultados

standardize country names and iso 2 iso3 codes

iso2ToIso3 para utilizar em coordinateCleaner

centroids

conferência textual
}
\section{0 - Preparar ambiente R}{
NA
}

\section{Selecionar UC}{
NA
}

\section{Install and load packeges to test}{
NA
}

\section{Gerar centroides}{
NA
}

\section{Preparaçao}{
NA
}

\section{Tela APP--}{
NA
}

\section{Server}{
NA
}

\section{padronizar nome de paises e codigos iso2 e iso3}{
NA
}

\section{coordinateCleaner and bdc}{
NA
}

\section{Run the application}{
NA
}

\examples{
\donttest{
app.prepare()
}
}
\seealso{
\code{\link[utils]{download.file}}, \code{\link[utils]{aspell}}
}
\author{
Pablo Hendrigo Alves de Melo,
Thuane Bochorny &
Rafaela Forzza
}
