% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_app.R
\name{app}
\alias{app}
\alias{load_app}
\title{Plant Catalog of Units of Brazilian Conservation App}
\usage{
load_app()
}
\value{
CSV files
}
\description{
An R app for preparing species listings for the Plant Catalog of Units of Brazilian Conservation.
}
\examples{
\donttest{
load_app()
}
}
\seealso{
\code{\link[utils]{download.file}}, \code{\link[utils]{aspell}}
}
\author{
Pablo Hendrigo Alves de Melo,
Thuane Bochorny &
Rafaela Forzza
}
