% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_publication.R
\name{app_publication}
\alias{app_publication}
\title{Plant Catalog of Units of Brazilian Conservation App}
\usage{
app_publication()
}
\value{
CSV files
}
\description{
An R app for preparing species listings for the Plant Catalog of Units of Brazilian Conservation.
}
\examples{
\donttest{
app_publication()
}
}
\seealso{
\code{\link[utils]{download.file}}, \code{\link[utils]{aspell}}
}
\author{
Pablo Hendrigo Alves de Melo,
Thuane Bochorny &
Rafaela Forzza
}
