% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_prepare.R
\name{app_prepare}
\alias{app_prepare}
\title{Prepare Plant Catalog of Units of Brazilian Conservation App}
\value{
CSV files
}
\description{
An R app for preparing species listings for the Plant Catalog of Units of Brazilian Conservation.
}
\examples{
\donttest{
app_prepare()
}

}
\seealso{
\code{\link[utils]{download.file}}, \code{\link[utils]{aspell}}
}
\author{
Pablo Hendrigo Alves de Melo,
Thuane Bochorny &
Rafaela Forzza
}
