% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_review - Copia.R, R/app_review.R
\name{app_review}
\alias{app_review}
\title{Plant Catalog of Units of Brazilian Conservation App}
\usage{
app_review()

app_review()
}
\value{
CSV files

CSV files
}
\description{
An R app for preparing species listings for the Plant Catalog of Units of Brazilian Conservation.

An R app for preparing species listings for the Plant Catalog of Units of Brazilian Conservation.
}
\examples{
\donttest{
app_review()
}
\donttest{
app_review()
}
}
\seealso{
\code{\link[utils]{download.file}}, \code{\link[utils]{aspell}}

\code{\link[utils]{download.file}}, \code{\link[utils]{aspell}}
}
\author{
Pablo Hendrigo Alves de Melo,
Thuane Bochorny &
Rafaela Forzza
}
